refactor(android): 알람 상수/룩업 object 를 AlarmConstants.kt 로 분리
AlarmEntity.kt(222줄)가 엔티티+드래프트와 11개 상수 object(AlarmStates,
VibrationPatterns, HolidaySeedData, AlarmPlayModes 등)를 함께 담고 있었음.
iOS AlarmEnums.swift 처럼 상수를 전용 파일로 분리(AlarmEntity.kt 90줄).
같은 패키지라 FQN 불변 → 참조처 수정 불필요. 동작 변경 없음.
